7주차 배포

+++ src/main/java/store/mybooks/resource/book/service/BookService.java
@@ -2,8 +2,8 @@ package store.mybooks.resource.book.service;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -11,6 +11,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import store.mybooks.resource.book.dto.request.BookCreateRequest;
@@ -22,6 +23,7 @@ import store.mybooks.resource.book.dto.response.BookDetailResponse;
 import store.mybooks.resource.book.dto.response.BookGetResponseForCoupon;
 import store.mybooks.resource.book.dto.response.BookModifyResponse;
 import store.mybooks.resource.book.dto.response.BookResponseForOrder;
+import store.mybooks.resource.book.dto.response.BookStockResponse;
 import store.mybooks.resource.book.entity.Book;
 import store.mybooks.resource.book.exception.BookNotExistException;
 import store.mybooks.resource.book.exception.IsbnAlreadyExistsException;
@@ -33,12 +35,12 @@ import store.mybooks.resource.bookcategory.dto.request.BookCategoryCreateRequest
 import store.mybooks.resource.bookcategory.service.BookCategoryService;
 import store.mybooks.resource.booklike.repository.BookLikeRepository;
 import store.mybooks.resource.bookstatus.entity.BookStatus;
+import store.mybooks.resource.bookstatus.enumeration.BookStatusEnum;
 import store.mybooks.resource.bookstatus.exception.BookStatusNotExistException;
 import store.mybooks.resource.bookstatus.respository.BookStatusRepository;
 import store.mybooks.resource.booktag.dto.request.BookTagCreateRequest;
 import store.mybooks.resource.booktag.service.BookTagService;
 import store.mybooks.resource.category.service.CategoryService;
-import store.mybooks.resource.image.dto.response.ImageRegisterResponse;
 import store.mybooks.resource.image.entity.Image;
 import store.mybooks.resource.image.exception.ImageNotExistsException;
 import store.mybooks.resource.image.repository.ImageRepository;
@@ -192,16 +194,15 @@ public class BookService {
             bookTagService.createBookTag(new BookTagCreateRequest(bookId, createRequest.getTagList()));
         }
 
-        List<ImageRegisterResponse> imageRegisterResponseList = new ArrayList<>();
         ImageStatus thumbnailEnum = imageStatusRepository.findById(ImageStatusEnum.THUMBNAIL.getName()).orElseThrow(
                 () -> new ImageStatusNotExistException("해당 하는 id의 이미지 상태가 없습니다."));
-        ImageRegisterResponse imageRegisterResponse = imageService.saveImage(thumbnailEnum, null, book, thumbnail);
-        imageRegisterResponseList.add(imageRegisterResponse);
+        imageService.saveImage(thumbnailEnum, null, book, thumbnail);
+
 
         ImageStatus contentEnum = imageStatusRepository.findById(ImageStatusEnum.CONTENT.getName())
                 .orElseThrow(() -> new ImageStatusNotExistException("해당 하는 id의 이미지 상태가 없습니다."));
         for (MultipartFile file : content) {
-            imageRegisterResponseList.add(imageService.saveImage(contentEnum, null, book, file));
+            imageService.saveImage(contentEnum, null, book, file);
         }
         return bookMapper.createResponse(newBook);
     }
@@ -217,16 +218,33 @@ public class BookService {
      * @return book modify response
      */
     @Transactional
-    public BookModifyResponse modifyBook(Long bookId, BookModifyRequest modifyRequest) {
+    public BookModifyResponse modifyBook(Long bookId, BookModifyRequest modifyRequest, MultipartFile thumbnail,
+                                         List<MultipartFile> content) throws IOException {
         Book book = bookRepository.findById(bookId)
                 .orElseThrow(() -> new BookNotExistException(bookId));
 
         BookStatus bookStatus = bookStatusRepository.findById(modifyRequest.getBookStatusId())
                 .orElseThrow(() -> new BookStatusNotExistException(modifyRequest.getBookStatusId()));
 
-        book.setModifyRequest(bookStatus, modifyRequest.getSaleCost(),
+        Publisher publisher =
+                publisherRepository.findById(modifyRequest.getPublisherId())
+                        .orElseThrow(() -> new PublisherNotExistException(modifyRequest.getPublisherId()));
+
+        if ((!Objects.equals(book.getIsbn(), modifyRequest.getIsbn()))
+                && (bookRepository.existsByIsbn(modifyRequest.getIsbn()))) {
+            throw new IsbnAlreadyExistsException(modifyRequest.getIsbn());
+        }
+
+        book.setModifyRequest(bookStatus, publisher, modifyRequest.getName(), modifyRequest.getIsbn(),
+                modifyRequest.getPublishDate(), modifyRequest.getPage(), modifyRequest.getIndex(),
+                modifyRequest.getExplanation(), modifyRequest.getOriginalCost(), modifyRequest.getSaleCost(),
                 ((book.getOriginalCost() - modifyRequest.getSaleCost()) * 100) / book.getOriginalCost(),
                 modifyRequest.getStock(), modifyRequest.getIsPacking());
+
+
+        bookAuthorService.deleteBookAuthor(bookId);
+        bookAuthorService.createBookAuthor(new BookAuthorCreateRequest(bookId, modifyRequest.getAuthorList()));
+
 
         bookCategoryService.deleteBookCategory(bookId);
         bookCategoryService.createBookCategory(new BookCategoryCreateRequest(bookId, modifyRequest.getCategoryList()));
@@ -235,13 +253,15 @@ public class BookService {
         if (modifyRequest.getTagList() != null) {
             bookTagService.createBookTag(new BookTagCreateRequest(bookId, modifyRequest.getTagList()));
         }
+        imageService.updateImage(book, thumbnail, content);
+
         return bookMapper.modifyResponse(book);
     }
 
     /**
      * methodName : getBookInCart
      * author : Fiat_lux
-     * description : 장바구니 안에 있는 책의 필요한 정보 가져오는 메서드
+     * description : 장바구니 안에 있는 책의 필요한 정보 가져오는 메서드.
      *
      * @param bookId the book id
      * @return BookCartResponse dto
@@ -251,10 +271,11 @@ public class BookService {
         Book book = bookRepository.findById(bookId).orElseThrow(() -> new BookNotExistException(bookId));
         Image image = imageRepository.findImageByBook_IdAndImageStatus_Id(bookId, ImageStatusEnum.THUMBNAIL.getName())
                 .orElseThrow(() -> new ImageNotExistsException("해당하는 id의 이미지가 없습니다"));
-
-
         String url = image.getPath() + image.getFileName() + image.getExtension();
-        return new BookCartResponse(book.getId(), book.getName(), url, book.getOriginalCost(), book.getSaleCost());
+        String bookStatusId = book.getBookStatus().getId();
+
+        return new BookCartResponse(book.getId(), book.getName(), url, book.getOriginalCost(), book.getSaleCost(),
+                book.getStock(), bookStatusId);
     }
 
     @Transactional(readOnly = true)
@@ -271,7 +292,7 @@ public class BookService {
     @Transactional
     public void updateBookViewCount() {
         Set<String> keys = redisTemplate.keys("viewCount:*");
-        if (keys != null && !keys.isEmpty()) {
+        if (!keys.isEmpty()) {
             for (String key : keys) {
                 Long bookId = Long.parseLong(key.substring("viewCount:".length()));
                 if (!bookRepository.existsById(bookId)) {
@@ -284,5 +305,41 @@ public class BookService {
                 }
             }
         }
+    }
+
+    /**
+     * methodName : getBookStockResponse<br>
+     * author : mins11<br>
+     * description : 책의 재고를 가지고 오는 메서드.<br>
+     *
+     * @param bookId the book id
+     * @return the book stock response
+     */
+    @Transactional(readOnly = true)
+    public BookStockResponse getBookStockResponse(Long bookId) {
+        return bookRepository.getBookStockList(bookId);
+    }
+
+    /**
+     * methodName : updateBookStock<br>
+     * author : mins11<br>
+     * description : 책 재고 업데이트.<br>
+     *
+     * @param bookId the book id
+     * @param stock  the stock
+     */
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    public void updateBookStock(Long bookId, Integer stock) {
+        Book book = bookRepository.findById(bookId)
+                .orElseThrow(() -> new BookNotExistException(bookId));
+        int resultStock = book.getStock() - stock;
+        if (resultStock == 0) {
+            BookStatus bookStatus = bookStatusRepository.findById(BookStatusEnum.NO_STOCK.getName())
+                    .orElseThrow(() -> new BookStatusNotExistException(BookStatusEnum.NO_STOCK.getName()));
+            book.soldOut(resultStock, bookStatus);
+        } else {
+            book.modifyStock(resultStock);
+        }
+
     }
 }
